refactor: moving main components to components/core
