+++ app/components/core/IncidentPopup.jsx
