+++ app/components/visualisations/ATTNetworkCollapse.jsx
