+++ app/components/visualisations/Terminal.jsx
