+++ app/components/visualisations/Therac25Accidents.jsx
