+++ app/components/visualisations/Win95RecycleBin.jsx
